Processor implementation, rather than export implementation
