+++ src/exportable/ExportableExtension.ts
@@ -1,7 +1,7 @@
 import { EOL } from 'os';
 import { fshtypes } from 'fsh-sushi';
 import { Exportable, ExportableSdRule } from '.';
-import { metadataToFSH, switchQuantityRules } from './common';
+import { metadataToFSH } from './common';
 
 export class ExportableExtension extends fshtypes.Extension implements Exportable {
   rules: ExportableSdRule[];
@@ -11,7 +11,6 @@ export class ExportableExtension extends fshtypes.Extension implements Exportabl
   }
 
   toFSH(): string {
-    switchQuantityRules(this);
     const metadataFSH = metadataToFSH(this);
     const rulesFSH = this.rules.map(r => r.toFSH()).join(EOL);
     return `${metadataFSH}${rulesFSH.length ? EOL + rulesFSH : ''}`;
