+++ src/exportable/ExportableInstance.ts
@@ -1,6 +1,6 @@
 import { EOL } from 'os';
 import { fshtypes } from 'fsh-sushi';
-import { metadataToFSH, switchQuantityRules } from './common';
+import { metadataToFSH } from './common';
 import { Exportable, ExportableAssignmentRule, ExportableInsertRule } from '.';
 
 export class ExportableInstance extends fshtypes.Instance implements Exportable {
@@ -11,7 +11,6 @@ export class ExportableInstance extends fshtypes.Instance implements Exportable 
   }
 
   toFSH(): string {
-    switchQuantityRules(this);
     const metadataFSH = metadataToFSH(this);
     const rulesFSH = this.rules.map(r => r.toFSH()).join(EOL);
     return `${metadataFSH}${rulesFSH.length ? EOL + rulesFSH : ''}`;
